add navbar
