+++ frontend/package.json
@@ -26,6 +26,6 @@
     "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^15.15.0",
-    "vite": "^6.3.4"
+    "vite": "^6.3.3"
   }
 }
