+++ frontend/src/Components/DocumentFolders.jsx
@@ -1,0 +1,5 @@
+export default function DocumentFolders() {
+    return (
+        <p>Tady bude vypis slozek</p>
+    )
+};
