+++ frontend/src/Components/SearchTool.jsx
@@ -1,0 +1,5 @@
+export default function SearchTool() {
+    return (
+        <p>Tady pujde delat prehledavani</p>
+    )
+};
